Core env files update. (#26)
* Core env files update.

* Added check for the default admin password for core.

Bumped core.

+++ deploy.sh
@@ -3,15 +3,20 @@
 # Move this file 1 level above the oms-docker installation!
 
 echo -e "###\n### Deploying OMS...\n###"
+if [ $(grep ADMIN_PASSWORD ./oms-core.env) == "ADMIN_PASSWORD=1234" ]; then
+  echo "ERROR: Please do not use the default ADMIN_PASSWORD in ./oms-core.env"
+  exit 9
+fi
 cd oms-docker/docker
 docker-compose down -v
 cd ../../
 rm -Rf oms-docker
 echo -e "\nFinished cleanup\n"
+
 git clone --recursive --branch dev https://github.com/AEGEE/oms-docker.git
-cp ./oms-core.env oms-docker/oms-core/example.env
-cp ./oms-core_UserSeeder.php oms-docker/oms-core/database/seeds/UserSeeder.php
+cp ./oms-core.env oms-docker/oms-core/.env.example
 echo -e "\nFinished setting up files\n"
+
 cd oms-docker/docker
 docker-compose up -d
 echo -e "\nInitializing..."
